Update retrieve and set information inside CrimeFragment
What is expected from a fragment is to be reusable, so instead to retrieve directly the information from the activity's intent, the Fragment now is using it's own Bundle that every Fragment has attached to it.

This is possible creating a new static method inside the Fragment named "FragmentRequired newInstance([args needed])" with this method the activity can call the method and pass the data to the Fragment's Bundle directly instead of calling Fragment's Constructor.