Fix update CrimeListFragment after edit any row
When the user changes any detail of the rows inside the CrimeListFragment and got back to it, the list didn't updaate the List.

The solution was notify a change to the Adapter and create a new instance of the Adapter and set the Adapter to the RecyclerView again when the user gets back from CrimeActivity.