Create values-sw600dp/refs.xml
This new file is going to be used when a device with a minimum width of 600dp is used thus is going to use activity_twopane.xml instead of using activiti_fragment.xml. In brief, this is going to determine which layout to use depending on if the device is a tablet or a smartphone.